Создал папку Helpers, перенёс в неё GetFromTxt и создал AdminList (выпилив из Constants/Id)
