+++ WeekChgkSPBParser/API/Connect.cs
@@ -1,6 +1,7 @@
 ﻿using Telegram.Bot;
 using Telegram.Bot.Extensions.Polling;
 using WeekChgkSPBParser.Constants;
+using WeekChgkSPBParser.Helpers;
 
 namespace WeekChgkSPBParser.API
 {
