+++ WeekChgkSPBParser/Constants/Commands.cs
@@ -4,7 +4,7 @@
     {
         internal const string announcement = "/announcement";
         internal const string announcementToChannel = "/announcementToChannel";
-        internal const string announcementFromTxt = "/announcementFromLj";
-        internal const string announcementFromTxtToChannel = "/announcementFromLjToChannel";
+        internal const string announcementFromLj = "/announcementFromLj";
+        internal const string announcementFromLjToChannel = "/announcementFromLjToChannel";
     }
 }
