+++ WeekChgkSPBParser/Constants/Paths.cs
@@ -2,7 +2,7 @@
 {
     internal record class Paths
     {
-        internal const string TxtAnnounce = @"anounsment.txt";
+        internal const string TxtAnnounce = @"announcement.txt";
         internal const string TgToken = @"TgToken.txt";
         internal const string AdminIds = @"IDs.txt";
         internal const string LJUrl = @"https://chgk-spb.livejournal.com/2596838.html";
